@@ -1,6 +1,6 @@
-import pytest
+# import pytest
+# import uvloop
 
-
-@pytest.fixture
-def anyio_backend() -> str:
-    return "asyncio"
+# @pytest.fixture(scope="session")
+# def event_loop_policy():
+#     return uvloop.EventLoopPolicy()
