@@ -15,7 +15,7 @@ def test_recursive_type_sdl() -> None:
     Verifies recursive type references appear in the generated SDL.
     """
     schema = gm.Schema(query=Node)
-    sdl = schema.sdl()
+    sdl = schema._core.as_sdl()
 
     assert "type Node" in sdl
     assert "children: [Node!]" in sdl
