@@ -3,114 +3,92 @@ from typing import AsyncIterator
 
 import pytest
 
+from grommet.context import Context
 from grommet.errors import GrommetTypeError
-from grommet.info import Info
 from grommet.metadata import TypeKind
 from grommet.resolver import (
     _is_asyncgen_callable,
     _is_coroutine_callable,
-    _normalize_info,
+    _resolver_arg_params,
     _resolver_name,
     _resolver_params,
     _wrap_resolver,
 )
 
 
-def test_normalize_info_variants() -> None:
-    """
-    Verifies info normalization accepts Info, mapping, and object variants.
-    """
-    info = Info(field_name="field", context="ctx", root="root")
-    assert _normalize_info(info) is info
-
-    info_dict = _normalize_info({"field_name": "name", "context": 1, "root": 2})
-    assert info_dict.field_name == "name"
-    assert info_dict.context == 1
-    assert info_dict.root == 2
-
-    class Dummy:
-        field_name = "dummy"
-        context = "ctx"
-        root = "root"
-
-    info_obj = _normalize_info(Dummy())
-    assert info_obj.field_name == "dummy"
-    assert info_obj.context == "ctx"
-    assert info_obj.root == "root"
-
-
 def test_wrap_resolver_missing_annotation_raises() -> None:
-    """
-    Ensures resolver wrapping fails when required annotations are missing.
-    """
+    """Ensures resolver wrapping fails when required annotations are missing."""
 
-    async def resolver(parent, info, value):  # type: ignore[no-untyped-def]
+    async def resolver(self, value):  # type: ignore[no-untyped-def]
         return value
 
     with pytest.raises(GrommetTypeError):
         _wrap_resolver(resolver, kind=TypeKind.OBJECT, field_name="value")
 
 
-@pytest.mark.anyio
-async def test_wrap_resolver_info_context_root() -> None:
-    """
-    Verifies wrapped resolvers receive coerced args and info/context/root.
-    """
+async def test_wrap_resolver_with_args() -> None:
+    """Verifies wrapped resolvers receive coerced args."""
 
-    async def resolver(parent, info, context, root, value: int) -> tuple:
-        return (parent, info.field_name, context, root, value)
+    async def resolver(self, value: int) -> tuple:
+        return (self, value)
 
     wrapper, arg_defs = _wrap_resolver(
         resolver, kind=TypeKind.OBJECT, field_name="value"
     )
     assert arg_defs[0]["name"] == "value"
 
-    result = await wrapper(
-        "parent", {"field_name": "f", "context": 1, "root": 2}, value="3"
+    result = await wrapper("parent", None, value="3")
+    assert result == ("parent", 3)
+
+
+async def test_wrap_resolver_with_context() -> None:
+    """Verifies wrapped resolvers receive Context when annotated."""
+
+    async def resolver(self, ctx: Context, value: int) -> tuple:
+        return (self, ctx, value)
+
+    wrapper, arg_defs = _wrap_resolver(
+        resolver, kind=TypeKind.OBJECT, field_name="value"
     )
-    assert result == ("parent", "f", 1, 2, 3)
+    assert arg_defs[0]["name"] == "value"
+
+    ctx_obj = object()
+    result = await wrapper("parent", ctx_obj, value="3")
+    assert result == ("parent", ctx_obj, 3)
 
 
-@pytest.mark.anyio
-async def test_wrap_resolver_without_info_and_missing_kwargs() -> None:
-    """
-    Verifies wrapped resolvers use defaults when info kwargs are omitted.
-    """
+async def test_wrap_resolver_without_args_uses_defaults() -> None:
+    """Verifies wrapped resolvers use defaults when kwargs are omitted."""
 
-    async def resolver(parent, value: int = 5) -> tuple:
-        return (parent, value)
+    async def resolver(self, value: int = 5) -> tuple:
+        return (self, value)
 
     wrapper, _ = _wrap_resolver(resolver, kind=TypeKind.OBJECT, field_name="value")
-    result = await wrapper("parent", {"field_name": "ignored"})
+    result = await wrapper("parent", None)
     assert result == ("parent", 5)
 
 
 def test_wrap_resolver_requires_async() -> None:
-    """
-    Ensures sync resolvers are rejected for non-subscription fields.
-    """
+    """Ensures sync resolvers are rejected for non-subscription fields."""
 
-    def resolver(parent, info) -> int:
+    def resolver(self) -> int:
         return 1
 
     with pytest.raises(GrommetTypeError):
         _wrap_resolver(resolver, kind=TypeKind.OBJECT, field_name="value")
 
 
-@pytest.mark.anyio
 async def test_wrap_subscription_requires_async_iterator() -> None:
-    """
-    Ensures subscription resolvers must return async iterators.
-    """
+    """Ensures subscription resolvers must return async iterators."""
 
-    async def resolver(parent, info) -> int:
+    async def resolver(self) -> int:
         return 3
 
     wrapper, _ = _wrap_resolver(
         resolver, kind=TypeKind.SUBSCRIPTION, field_name="ticks"
     )
     with pytest.raises(GrommetTypeError):
-        await wrapper("parent", {"field_name": "ticks"})
+        await wrapper(None, None)
 
 
 def test_resolver_params_handles_unhashable_callable() -> None:
@@ -126,6 +104,27 @@ def test_resolver_params_handles_unhashable_callable() -> None:
     params = _resolver_params(callable_obj)
     assert len(params) == 1
     assert params[0].name == "x"
+
+
+def test_resolver_arg_params_skips_self_and_context() -> None:
+    """Verifies _resolver_arg_params excludes self and Context parameters."""
+
+    async def resolver(self, ctx: Context[dict], value: int) -> int:
+        return value
+
+    params = _resolver_arg_params(resolver)
+    assert len(params) == 1
+    assert params[0].name == "value"
+
+
+def test_resolver_arg_params_self_only() -> None:
+    """Verifies _resolver_arg_params returns empty for self-only resolvers."""
+
+    async def resolver(self) -> int:
+        return 1
+
+    params = _resolver_arg_params(resolver)
+    assert len(params) == 0
 
 
 def test_resolver_name_fallback_to_type_name() -> None:
@@ -181,7 +180,7 @@ def test_is_asyncgen_callable_with_func_attr() -> None:
 def test_wrap_resolver_subscription_asyncgen_path() -> None:
     """Test _wrap_resolver handles subscription with async generator."""
 
-    async def subscription_gen(parent: None) -> AsyncIterator[int]:
+    async def subscription_gen(self) -> AsyncIterator[int]:
         yield 1
         yield 2
 
@@ -194,7 +193,7 @@ def test_wrap_resolver_subscription_asyncgen_path() -> None:
 def test_wrap_resolver_subscription_coroutine_path() -> None:
     """Test _wrap_resolver handles subscription with coroutine returning iterator."""
 
-    async def subscription_coro(parent: None) -> AsyncIterator[int]:
+    async def subscription_coro(self) -> AsyncIterator[int]:
         async def gen() -> AsyncIterator[int]:
             yield 1
 
@@ -206,11 +205,10 @@ def test_wrap_resolver_subscription_coroutine_path() -> None:
     assert arg_defs == []
 
 
-@pytest.mark.anyio
 async def test_wrap_resolver_subscription_coroutine_returns_iterator() -> None:
     """Verifies subscription coroutine wrapper awaits and returns the iterator."""
 
-    async def subscription_coro(parent: None) -> AsyncIterator[int]:
+    async def subscription_coro(self) -> AsyncIterator[int]:
         async def gen() -> AsyncIterator[int]:
             yield 42
 
@@ -219,14 +217,14 @@ async def test_wrap_resolver_subscription_coroutine_returns_iterator() -> None:
     wrapper, _ = _wrap_resolver(
         subscription_coro, kind=TypeKind.SUBSCRIPTION, field_name="events"
     )
-    result = await wrapper(None, {"field_name": "events"})
+    result = await wrapper(None, None)
     assert hasattr(result, "__anext__")
 
 
 def test_wrap_resolver_sync_subscription_rejected() -> None:
     """Ensures sync functions are rejected as subscription resolvers."""
 
-    def sync_sub(parent: None) -> int:  # type: ignore[return]
+    def sync_sub(self) -> int:  # type: ignore[return]
         return 1
 
     with pytest.raises(GrommetTypeError):
