@@ -1,25 +1,17 @@
 import asyncio
 from dataclasses import dataclass
-from typing import TYPE_CHECKING
-
-import pytest
 
 import grommet as gm
-
-if TYPE_CHECKING:
-    from typing import Any
 
 
 @gm.type
 @dataclass
 class Query:
     @gm.field
-    @staticmethod
-    async def ping(parent: "Any", info: "Any") -> str:
+    async def ping(self) -> str:
         return "pong"
 
 
-@pytest.mark.anyio
 async def test_configure_runtime_allows_execution() -> None:
     """
     Verifies runtime configuration enables schema execution.
@@ -28,10 +20,9 @@ async def test_configure_runtime_allows_execution() -> None:
     schema = gm.Schema(query=Query)
     result = await schema.execute("{ ping }")
 
-    assert result["data"]["ping"] == "pong"
+    assert result.data["ping"] == "pong"
 
 
-@pytest.mark.anyio
 async def test_nested_event_loop_execution() -> None:
     """
     Ensures schema execution works within nested async scheduling.
@@ -40,7 +31,7 @@ async def test_nested_event_loop_execution() -> None:
 
     async def run_query() -> str:
         payload = await schema.execute("{ ping }")
-        value = payload["data"]["ping"]
+        value = payload.data["ping"]
         assert isinstance(value, str)
         return value
 
