@@ -1,12 +1,6 @@
 from dataclasses import dataclass
-from typing import TYPE_CHECKING
-
-import pytest
 
 import grommet as gm
-
-if TYPE_CHECKING:
-    from typing import Any
 
 
 @gm.type
@@ -27,18 +21,16 @@ class SdlUserInput:
 @dataclass
 class SdlQuery:
     @gm.field
-    @staticmethod
-    async def greet(parent: "Any", info: "Any", user: SdlUserInput) -> str:
+    async def greet(self, user: SdlUserInput) -> str:
         return f"hi {user.name}"
 
 
-@pytest.mark.anyio
 async def test_schema_sdl_contains_types() -> None:
     """
     Verifies generated SDL includes referenced types and omits unused ones.
     """
     schema = gm.Schema(query=SdlQuery)
-    sdl = schema.sdl()
+    sdl = schema._core.as_sdl()
 
     assert "type Query" in sdl
     assert "input UserInput" in sdl
